requirements.txt
aiogram==2.25.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
apscheduler==3.10.4
matplotlib==3.8.2
python-dotenv==1.0.0
aiohttp==3.9.1